Refactor direct controller for reuse
